Move existing javascripts under webpacker
