+++ app/javascript/actions/players.js
