+++ app/javascript/actions/table.js
