+++ app/javascript/api/v2/fetch.js
