+++ app/javascript/api/v2/players.js
