+++ app/javascript/api/v2/table.js
