+++ app/javascript/banner.js
