+++ app/javascript/cable.js
