+++ app/javascript/components/scoreboard/ActivePlayersList.js
