+++ app/javascript/components/scoreboard/App.js
