+++ app/javascript/components/scoreboard/InactivePlayersList.js
