+++ app/javascript/components/scoreboard/MatchResult.js
