+++ app/javascript/components/scoreboard/Scoreboard.js
