+++ app/javascript/components/scoreboard/TableControls.js
