+++ app/javascript/packs/default.js
@@ -1,0 +1,1 @@
+import 'banner'
