+++ app/javascript/reducers/ongoingMatch.js
