+++ app/javascript/reducers/players.js
