+++ app/javascript/reducers/table.js
