+++ app/javascript/selectors/table.js
