+++ app/javascript/stores/scoreboard.js
