+++ app/javascript/utils/createStore.js
