+++ app/javascript/utils/csrf.js
