+++ app/javascript/utils/formatTime.js
