--- app/webpack/default.js
@@ -1,1 +1,0 @@
-import './banner'
